FINALLY! Notifications works for 'taken' action

+++ MyMeds/NotificationScheduler.h
@@ -24,6 +24,6 @@
 + (void)removeNotif:(int)number andDay:(NSString *)day andTime:(NSString*)timeString andDayIndex:(int)dayIndex;
 
 + (void)alterNotificationsForTakenMed:(TodayMedication*)med;
-+ (void)scheduleTodayNotificationsWithNumber:(int)number AndDay:(NSString*)day AndDayIndex:(int)dayIndex AndTime:(int)time;
++ (void)scheduleTodayNotificationsWithNumber:(int)number AndDay:(NSString*)day AndDayIndex:(int)dayIndex AndTime:(float)time;
 
 @end
